% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_management.R
\name{last_post}
\alias{last_post}
\alias{post_stack}
\alias{rm_last_post}
\title{Manage Posts}
\usage{
last_post()

post_stack()

rm_last_post()
}
\description{
Query/Remove posts created by slackreprex. Each post is logged
for channel and timestamp (ts) in an internal package environment. Which
serves the purpose of querying or removing specific posts.
\itemize{
\item post_stack: lists in decreasing order the posts sent during the session.
\item last_post: retrieves the last post sent.
\item rm_last_post: deletes the last post from the specified channel.
}
}
\seealso{
\link[slackcalls:delete_post]{delete_post}
}
