% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessioninfo.R
\name{post_sessioninfo}
\alias{post_sessioninfo}
\title{Post session info to Slack}
\usage{
post_sessioninfo(
  channel,
  ts = NULL,
  initial_comment = NULL,
  ...,
  token = Sys.getenv("SLACK_API_TOKEN")
)
}
\arguments{
\item{channel}{character, channel to post to}

\item{ts}{character, Slack API time stamp, Default: NULL}

\item{initial_comment}{character, comment that precedes the snippet, Default: NULL}

\item{\dots}{arguments passed to sessioninfo::session_info() if installed else passed to sessionInfo()}

\item{token}{Slack API token, Default: Sys.getenv("SLACK_API_TOKEN")}
}
\value{
\link[httr:response]{response} from Slack api
}
\description{
Post sessioninfo to Slack as a snippet
}
\details{
\code{channel} can be on of the following in an activated team
\itemize{
\item name of channel
\item channel id
\item link to a message in a channel, direct message or a group message
}

A pairing of \code{channel} and \code{ts} can be used instead of a link to post to a thread.
}
\examples{
\dontrun{
if(interactive()){

# assumes '~/.slackteams' exists.
slackteams::load_team()
team <- slackteams::get_teams()[1]
slackteams::activate_team(team)

post_sessioninfo(channel = 'slackbot')

 }
}
}
