% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprex_to_blocks.R
\name{reprex_to_blocks}
\alias{reprex_to_blocks}
\title{FUNCTION_TITLE}
\usage{
reprex_to_blocks(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
