% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprex.R
\name{slack_reprex}
\alias{slack_reprex}
\title{FUNCTION_TITLE}
\usage{
slack_reprex(..., text = NULL, channel, token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{...}{PARAM_DESCRIPTION}

\item{text}{PARAM_DESCRIPTION}

\item{channel}{PARAM_DESCRIPTION}

\item{token}{PARAM_DESCRIPTION, Default: Sys.getenv("SLACK_API_TOKEN")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[reprex]{reprex}}
}
