% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprex.R
\name{slack_reprex}
\alias{slack_reprex}
\title{post reprex output to slack}
\usage{
slack_reprex(
  ...,
  text = NULL,
  channel = NULL,
  ts = NULL,
  token = Sys.getenv("SLACK_API_TOKEN")
)
}
\arguments{
\item{...}{code to evaluate in reprex}

\item{text}{question text to attach to the reprex blocks}

\item{channel}{character, channel to post to}

\item{ts}{character, Slack API time stamp, Default: NULL}

\item{token}{Slack API token, Default: Sys.getenv("SLACK_API_TOKEN")}
}
\value{
\link[httr:response]{response} from slack api
}
\description{
Converts reprex output to slackblocks and
posts the blocks to a slack conversation which can be a channel,
a direct message, a group message or a thread in one of them.
}
\details{
\code{channel} can be on of the following in an activated team
\itemize{
\item name of channel
\item channel id
\item link to a message in a channel, direct message or a group message
}

A pairing of \code{channel} and \code{ts} can be used instead of a link to post to a thread.
}
\examples{
\dontrun{
if(interactive()){

slack_reprex({
  x <- 10
  hist(runif(x))
  hist(runif(2*y))
},
text = 'My question is ... ',
channel = 'slackbot')
 }
}
}
\seealso{
\link[reprex:reprex]{reprex}
}
